Split NodeDetails into separate components for each section
